ajustes tributação ncm

+++ Escon.SisctNET.Web/Controllers/CompanyController.cs
@@ -374,8 +374,8 @@ namespace Escon.SisctNET.Web.Controllers
                 rst.IcmsNContribuinte = entity.IcmsNContribuinte;
                 rst.IcmsNContribuinteFora = entity.IcmsNContribuinteFora;
                 rst.IcmsAliqM25 = entity.IcmsAliqM25;
-                rst.ChapterId = entity.ChapterId.Equals(0) ? null : entity.ChapterId;
-                rst.SectionId = entity.SectionId.Equals(0) ? null : entity.SectionId;
+                rst.ChapterId = entity.ChapterId.Equals((long)0) ? null : entity.ChapterId;
+                rst.SectionId = entity.SectionId.Equals((long)0) ? null : entity.SectionId;
                 rst.AliqInterna = entity.AliqInterna;
                 rst.IncIInterna = entity.IncIInterna;
                 rst.IncIInterestadual = entity.IncIInterestadual;
